mode make_type to test mod
