+++ cli/src/utils.rs
@@ -94,13 +94,3 @@ pub fn with_indent(s: String, indent: usize) -> String {
         .collect::<Vec<_>>()
         .join("\n")
 }
-
-/// Given a type definition, return type ID and registry representing it.
-#[allow(dead_code)]
-pub fn make_type<T: scale_info::TypeInfo + 'static>() -> (u32, scale_info::PortableRegistry) {
-    let m = scale_info::MetaType::new::<T>();
-    let mut types = scale_info::Registry::new();
-    let id = types.register_type(&m);
-    let portable_registry: scale_info::PortableRegistry = types.into();
-    (id.id, portable_registry)
-}
